CR Feedback - reorg test code - remove debug garbage - demo app and catalog should respect unique identifiers
