+++ Source/List/ListSection.swift
@@ -24,8 +24,8 @@ public struct ListSection {
     /// Identical to proportionally except that the last cell will take up the remaining width.
     /// If `minimumWidth` is not available, it will be full width on a new line.
     /// This is useful when, for example, displaying an input field at the end of an array of items.
-    /// Your last cellModel should return `.relative` from  `size(constrainedTo:)` to
-    /// get fill-remaining-width behavior. 
+    /// Your last cellModel should return `.relative` from  `size(constrainedTo:)` to get
+    /// fill-remaining-width behavior, or  `.autolayout` to keep `.proportionally` behavior.
     case proportionallyWithLastCellFillingWidth(minimumWidth: CGFloat)
   }
